Backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
faiss-cpu>=1.9.0
python-dotenv==1.0.0
pypdf>=4.0.0
pydantic>=2.5.0
requests==2.31.0
numpy>=1.24.0
scikit-learn>=1.3.0
